requirements.txt
anyio==3.7.0
attrs==23.1.0
Automat==22.10.0
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
cryptography==41.0.1
cssselect==1.2.0
dnspython==2.3.0
fastapi==0.98.0
filelock==3.12.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jmespath==1.0.1
joblib==1.3.0
lxml==4.9.2
nltk==3.8.1
packaging==23.1
parsel==1.8.1
Protego==0.2.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.9
PyDispatcher==2.0.7
pymongo==4.4.0
pyOpenSSL==23.2.0
python-dotenv==1.0.0
queuelib==1.6.2
regex==2023.6.3
requests==2.31.0
requests-file==1.5.1
rfc3986==1.5.0
Scrapy==2.9.0
service-identity==23.1.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
starlette==0.27.0
textblob==0.17.1
tldextract==3.4.4
tqdm==4.65.0
Twisted==22.10.0
typing_extensions==4.6.3
urllib3==2.0.3
uvicorn==0.22.0
w3lib==2.1.1
zope.interface==6.0
